Fix lint warning
